verizon_chatbot/requirements.txt
anthropic==0.8.0
beautifulsoup4==4.12.2
fastapi==0.104.1
python-dotenv==1.0.0
requests==2.31.0
uvicorn==0.24.0
pydantic==2.4.2
selenium==4.15.2
python-multipart==0.0.6
pydantic_settings